src/third_party/blink/web_tests/xr/xrViewport_valid.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/device/vr/public/mojom/vr_service.mojom.js"></script>
<script src="../external/wpt/resources/chromium/webxr-test.js"></script>
<script src="../external/wpt/webxr/resources/webxr_test_constants.js"></script>
<script src="../xr/resources/xr-test-utils.js"></script>
<canvas id="webgl-canvas"></canvas>

<script>
let testName = "XRViewport attributes are valid";

let fakeDeviceInitParams = { supportsImmersive:true };

let requestSessionOptions = [{ mode: 'immersive-vr' }];

let testFunction = function(session, t, fakeDeviceController) {
  // Session must have a baseLayer or frame requests will be ignored.
  let webglLayer = new XRWebGLLayer(session, gl);
  session.updateRenderState({ baseLayer: webglLayer });

  // Need to have a valid pose or input event's don't process.
  fakeDeviceController.setXRPresentationFrameData(VALID_POSE_MATRIX, [{
      eye:"left",
      projectionMatrix: VALID_PROJECTION_MATRIX,
      viewMatrix: VALID_VIEW_MATRIX
    }, {
      eye:"right",
      projectionMatrix: VALID_PROJECTION_MATRIX,
      viewMatrix: VALID_VIEW_MATRIX
    }]);

  return session.requestReferenceSpace({ type: "stationary", subtype: "eye-level" })
    .then((referenceSpace) => new Promise((resolve) =>{
      function onFrame(time, xrFrame) {
        let pose = xrFrame.getViewerPose(referenceSpace);
        console.log(pose);
        console.log(pose.views);
        assert_not_equals(pose, null);
        assert_not_equals(pose.views, null);
        assert_equals(pose.views.length, 2);

        let leftView = pose.views[0];
        let rightView = pose.views[1];

        let leftViewport = webglLayer.getViewport(leftView);
        let rightViewport = webglLayer.getViewport(rightView);

        t.step(() => {
          // Ensure the views report the expected viewports into the WebGL layer.
          assert_true(leftViewport instanceof XRViewport);
          assert_true(rightViewport instanceof XRViewport);

          assert_not_equals(leftViewport, null);
          assert_not_equals(rightViewport, null);

          // Exact viewport values don't matter, but they must pass several tests:

          // Viewports have non-zero widths and heights.
          assert_greater_than(leftViewport.width, 0);
          assert_greater_than(leftViewport.height, 0);
          assert_greater_than(rightViewport.width, 0);
          assert_greater_than(rightViewport.height, 0);

          // Viewports are located within the framebuffer.
          assert_greater_than_equal(leftViewport.x, 0);
          assert_greater_than_equal(leftViewport.y, 0);
          assert_greater_than_equal(leftViewport.x, 0);
          assert_greater_than_equal(leftViewport.y, 0);

          assert_less_than_equal(
            leftViewport.x + leftViewport.width, webglLayer.framebufferWidth);
          assert_less_than_equal(
            leftViewport.y + leftViewport.height, webglLayer.framebufferHeight);
          assert_less_than_equal(
            rightViewport.x + rightViewport.width, webglLayer.framebufferWidth);
          assert_less_than_equal(
            rightViewport.y + rightViewport.height, webglLayer.framebufferHeight);

          // Assumes that the left viewport will always be physically to the left
          // of the right viewport.
          assert_less_than_equal(leftViewport.x + leftViewport.width, rightViewport.x);
        });

        resolve();
      }
      session.requestAnimationFrame(onFrame);
  }));
};

xr_session_promise_test(
  testFunction, fakeDeviceInitParams, requestSessionOptions, testName);

</script>
